fix(789): ensure children passed to background component update
This commit refactors the `shouldComponentUpdate` function outside of
the `BackgroundImpl` component.

The refactored function, `__shouldComponentUpdate` takes `props` and
`nextProps` as arguments and returns true if the children, imgixParams,
contextRect, bounds, htmlAttributes, or top-level props have changed.
It returns false if the props have not changed.

The refactoring addresses an issue where changes to the children and
imgixParams props were not detected and the child components therefore
did not re-render.

Co-authored-by: Frederick Fogerty <frederick@imgix.com>