fix: move consumer validation to avoid unnecessary warnings
This commit moves the block that ensured the Provider was being used
inside the appropriate context into the component definition. This
removes the possibility of it logging a warning when ReactImgix
components are wrapped without the context component while also issuing
a warning when the provider is used without consumers.