refactor: remove hasDomain function and use indexOf instead
This commit removed the `hasDomain()` function. It was too unreliable at
correctly determining when a URL had a domain prefix. Instead,
requiring that `src` be formatted with a domain prefix when using a
1-step URL is both a more maintainable and easier to explain solution.
Moreover, it matches existing behavior nicely.

In future, we might want to add validations so that a source with
incomplete domain information, ie, "foo.imgix.net/bar.jpg" at least
results in some sort of warning.