ci: fix node:current tests
CircleCI's latest node (v18.0.0) orb is broken; more specifically it does not install node or npm. For now, the best solution is to pin the `current` tests against the latest working orb (v17.9). We should check back on this in the future and revert to point at the latest version once fixed.

Co-authored-by: Luis Ball <lball@imgix.com>