feat: allow width and height to be passed through htmlAttributes
[It's best practice to specify width and height on images][1], even if
they are fluid, to help avoid layout shift. It tells the browser the
expected size before it loads any image data, and an aspect ratio is
derived from it so the browser can still know the image size even if CSS
or something else is altering the image's width from its intrinsic one.

Before this patch, there was not a way to allow width and height
attributes to get through to the `img` element sent to the browser when
using the react-imgix component while leaving the image fluid width
(that is, with a srcset containing width descriptors). As soon as one
specified width and height, react-imgix assumed the author wanted a
fixed-width image, and it gave resolution descriptors in the srcset
instead of width descriptors (and then the sizes attribute was ignored
by the browser, [as per spec][2]).

Imgix currently uses the width and height attributes to decide whether
to use width or resolution descriptors, so to avoid altering that logic
(which would be a breaking change), this patch allows width and height
to be passed through the `htmlAttributes` prop. Prior to the patch they
were swallowed if given in `htmlAttributes`. As of this patch they are
allowed and passed through.

This closes https://github.com/imgix/react-imgix/issues/891

In future it may make sense to decide srcset type based on the presence
of the `sizes` attribute, as discussed in the above ticket. This would
be a breaking change, however.

As part of this patch, width and height are also allowed through when
using this component in `Source` mode, since [width and height
attributes are now allowed and encouraged in that context][3].

[1]: https://web.dev/optimize-cls/
[2]: https://html.spec.whatwg.org/multipage/images.html#srcset-attributes
[3]: https://github.com/whatwg/html/pull/5894